feat: hitting jump+carry when carrying a spring gets the extra jump height from the spring
