--- src/game/physics/mechanics/carrying.ts
@@ -1,238 +1,0 @@
-import type { UnionOfAllItemInPlayTypes } from "../../../model/ItemInPlay";
-import type { HeelsAbilities } from "../../../model/ItemStateMap";
-import type { Collideable } from "../../collision/aabbCollision";
-import type { GameState } from "../../gameState/GameState";
-import type { GridSpatialIndex } from "../gridSpace/GridSpatialIndex";
-import type { PlayableItem, PortableItem } from "../itemPredicates";
-
-import {
-  iterateRoomItems,
-  roomSpatialIndexKey,
-  type RoomState,
-} from "../../../model/RoomState";
-import { getEffectivelyStandingOnItemIdForPlayable } from "../../../model/stoodOnItemsLookup";
-import { blockSizePx } from "../../../sprites/spritePivots";
-import { addXyz } from "../../../utils/vectors/vectors";
-import { collisionItemWithIndex } from "../../collision/aabbCollision";
-import { findStandingOnWithHighestPriorityAndMostOverlap } from "../../collision/checkStandingOn";
-import { playableHasShield } from "../../gameState/gameStateSelectors/selectPickupAbilities";
-import { addItemToRoom } from "../../gameState/mutators/addItemToRoom";
-import { deleteItemFromRoom } from "../../gameState/mutators/deleteItemFromRoom";
-import { removeStandingOn } from "../../gameState/mutators/standingOn/removeStandingOn";
-import { handleItemsTouchingItems } from "../handleTouch/handleItemsTouchingItems";
-import { isDeadly, isPortable, isSolid } from "../itemPredicates";
-import { isFreeItem } from "../itemPredicates";
-import { moveItem } from "../moveItem/moveItem";
-
-/**
- * After pressing handling carry being pressed, how long until the action repeats?
- * This allows actions such as jump+carry to pick an item up, holding that button,
- * and immediately jump+carrying again on landing
- */
-const inputLatchDuration = 350;
-
-/**
- * walking, but also gliding and changing direction mid-air
- */
-export const carrying = <RoomId extends string, RoomItemId extends string>(
-  carrier: PlayableItem<"headOverHeels" | "heels", RoomId, RoomItemId>,
-  room: RoomState<RoomId, RoomItemId>,
-  gameState: GameState<RoomId>,
-  deltaMS: number,
-): undefined => {
-  const { inputStateTracker } = gameState;
-
-  const heelsAbilities =
-    carrier.type === "heels" ? carrier.state : carrier.state.heels;
-
-  const { carrying, hasBag } = heelsAbilities;
-  const {
-    state: { position: carrierPosition },
-  } = carrier;
-
-  if (!hasBag) {
-    return;
-  }
-
-  const portableRoomItemsIter = iterateRoomItems(room.items).filter(isPortable);
-  const itemToPickup =
-    carrying === null ? findItemToPickup(carrier, room) : undefined;
-
-  // update marking items as the next to pick up, for the sake of the green outline:
-  for (const portableItem of portableRoomItemsIter) {
-    portableItem.state.wouldPickUpNext = false;
-  }
-  if (itemToPickup !== undefined) itemToPickup.state.wouldPickUpNext = true;
-
-  const currentCarryPress = inputStateTracker.currentActionPress("carry");
-  const hasCarryInput = currentCarryPress !== "released";
-
-  if (hasCarryInput) {
-    if (carrying === null) {
-      // trying to pick up
-      if (itemToPickup === undefined) {
-        // nothing to pick up
-        return;
-      }
-
-      pickUpItem(room, heelsAbilities, itemToPickup);
-
-      // won't carry again until key is released and re-pressed - prevents
-      // multiple pickup/putdown in one tick with multiple sub-ticks
-      inputStateTracker.inputWasHandled("carry", inputLatchDuration);
-    } else {
-      // trying to put down
-      if (carrier.state.standingOnItemId === null) {
-        // can't put down mid-air
-        return;
-      }
-
-      // check if there is space above heels (and any items standing on heels):
-      // really the ideal here would be do the move and roll back if ti can't be done.
-      // that would need a stateless/reducer based approach to updating the world
-      if (!checkSpaceAvailableToPutDown(carrier, room[roomSpatialIndexKey])) {
-        return;
-      }
-
-      addItemToRoom({
-        room,
-        item: carrying,
-        atPosition: carrierPosition,
-      });
-
-      // ⬇ player isn't standing on whatever they were standing on before
-      removeStandingOn(carrier, room);
-
-      // not carrying it any more
-      heelsAbilities.carrying = null;
-
-      // moveItem could cause collision with a reincarnation fish, so it must be called only
-      // once the room is back into a good state (eg, carrying was set to null) or the
-      // item being put down could be duplicated
-      moveItem({
-        subjectItem: carrier,
-        gameState,
-        room,
-        posDelta: {
-          x: 0,
-          y: 0,
-          z: carrying.aabb.z,
-        },
-        forceful: true,
-        deltaMS,
-        onTouch: handleItemsTouchingItems,
-        path: new Set([carrier.id]),
-      });
-
-      // don't set heels as standing on the put-down item - normal gravity and movement
-      // will sort that out from the main loop
-      inputStateTracker.inputWasHandled("carry", inputLatchDuration);
-    }
-  }
-};
-
-const pickUpItem = <RoomId extends string, RoomItemId extends string>(
-  room: RoomState<RoomId, RoomItemId>,
-  heelsAbilities: HeelsAbilities<RoomId>,
-  itemToCarry: PortableItem<RoomId, RoomItemId>,
-) => {
-  heelsAbilities.carrying = itemToCarry;
-
-  itemToCarry.state.wouldPickUpNext = false;
-  deleteItemFromRoom({ room, item: itemToCarry });
-};
-
-export const findItemToPickup = <
-  RoomId extends string,
-  RoomItemId extends string,
->(
-  carrier: PlayableItem<"headOverHeels" | "heels", RoomId, RoomItemId>,
-  room: RoomState<RoomId, RoomItemId>,
-) => {
-  const hasShield = playableHasShield(carrier);
-
-  const itemIsPortableForCarrier = (
-    i: UnionOfAllItemInPlayTypes<RoomId, RoomItemId>,
-  ): i is PortableItem<RoomId, RoomItemId> =>
-    isPortable(i) &&
-    // can only pick up deadly items if you have a shield:
-    (hasShield || !isDeadly(i));
-
-  const portableItemsIter = iterateRoomItems(room.items).filter(
-    itemIsPortableForCarrier,
-  );
-
-  const straightStoodOn = findStandingOnWithHighestPriorityAndMostOverlap(
-    carrier,
-    portableItemsIter,
-  );
-
-  if (straightStoodOn) {
-    return straightStoodOn;
-  }
-
-  // nothing straight-up stood on, let's check if we're standing on by coyote time:
-  const coyoteStoodOnItemId = getEffectivelyStandingOnItemIdForPlayable(
-    room,
-    carrier.state,
-  );
-  const coyoteStoodOn = coyoteStoodOnItemId && room.items[coyoteStoodOnItemId];
-
-  if (coyoteStoodOn && itemIsPortableForCarrier(coyoteStoodOn)) {
-    return coyoteStoodOn;
-  }
-
-  return undefined;
-};
-
-export const checkSpaceAvailableToPutDown = <
-  T extends UnionOfAllItemInPlayTypes,
->(
-  item: T,
-  roomSpatialIndex: GridSpatialIndex,
-) => {
-  const proposedNewLocation: Collideable = {
-    state: {
-      position: addXyz(item.state.position, { z: blockSizePx.h }),
-    },
-    aabb: item.aabb,
-    id: `item.id-proposedPutdownLocation`,
-  };
-
-  const collisions = collisionItemWithIndex(
-    proposedNewLocation,
-    roomSpatialIndex,
-    // only check for collisions with solid items
-    (otherItem) =>
-      isSolid(otherItem, item) &&
-      // while in symbiosis, a proposed space one block higher can collide
-      // the the character doing the proposing - skip that:
-      otherItem !== item,
-  );
-
-  for (const collision of collisions) {
-    if (!isFreeItem(collision)) {
-      console.log(
-        "carrying: cannot put down due to collision: item:",
-        item,
-        "can't move up because it would collide with non-free",
-        collision,
-      );
-      return false;
-    }
-
-    // if there is a collision, check if it can be moved up too:
-    if (!checkSpaceAvailableToPutDown(collision, roomSpatialIndex)) {
-      console.log(
-        "carrying: cannot put down due to collision: item:",
-        item,
-        "can't move up because it would collide with free that has nowhere to go:",
-        collision,
-      );
-
-      return false;
-    }
-  }
-
-  return true;
-};
